countLeaves function
const newCount = count + 1 would just be counting 1 every time we go to the next height.
In the end, we are just returning the Max height between the left and the right subtree.
--- 
I suggest using only `count`, we add `1` once we reach the point where the node has no left and right child. And return the sum of both left and right count.